Move StyleScrollbarState to its own file https://bugs.webkit.org/show_bug.cgi?id=268245 rdar://121769126
Reviewed by Aditya Keerthi.

Having this be part of SelectorChecker.h causes circular imports with the change from bug 268063.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/SelectorChecker.h:
* Source/WebCore/style/PseudoElementRequest.h: Drive-by fix the license header
* Source/WebCore/style/StyleScrollbarState.h: Added.

Canonical link: https://commits.webkit.org/273617@main
