+++ Source/WebCore/css/SelectorChecker.h
@@ -32,6 +32,7 @@
 #include "SelectorMatchingState.h"
 #include "StyleRelations.h"
 #include "StyleScopeOrdinal.h"
+#include "StyleScrollbarState.h"
 
 namespace WebCore {
 
@@ -39,16 +40,6 @@ class CSSSelector;
 class Element;
 class RenderScrollbar;
 class RenderStyle;
-
-struct StyleScrollbarState {
-    ScrollbarPart scrollbarPart { NoPart };
-    ScrollbarPart hoveredPart { NoPart };
-    ScrollbarPart pressedPart { NoPart };
-    ScrollbarOrientation orientation { ScrollbarOrientation::Vertical };
-    ScrollbarButtonsPlacement buttonsPlacement { ScrollbarButtonsNone };
-    bool enabled { false };
-    bool scrollCornerIsVisible { false };
-};
 
 class SelectorChecker {
     WTF_MAKE_NONCOPYABLE(SelectorChecker);
